requirements.txt
psutil~=5.9.5
pyusb~=1.2.1
